Change fromMap to accept int input in destinationMdl

+++ lib/models/destinations/destination_mdl.dart
@@ -89,8 +89,8 @@ class DestinationMdl extends Equatable {
       province:
           map['province'] != null ? ProvinceMdl.fromMap(map['province']) : null,
       invitationsCount: map['invitations_count'],
-      latitude: map['latitude'],
-      longitude: map['longitude'],
+      latitude: map['latitude'].toDouble(),
+      longitude: map['longitude'].toDouble(),
     );
   }
